% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s2sls.R
\name{s2sls.formula}
\alias{s2sls.formula}
\title{formula}
\usage{
\method{s2sls}{formula}(formula, data = list(), w, ...)
}
\arguments{
\item{formula}{log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp}

\item{data}{the dataframe}

\item{w}{is the contiguity matrix}

\item{...}{not used}
}
\description{
formula
}

