% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s2sls.R
\name{s2sls}
\alias{s2sls}
\title{method}
\usage{
s2sls(x, ...)
}
\arguments{
\item{x}{a numeric design matrix for the model.}

\item{...}{not used}
}
\description{
method
}
\author{
Zaghdoudi Taha
}

